Update ConnectSchema.java
Bug fix. I'm passing in a SchemaBuilder for schema, and this will always throw an exception since it's doing an equals comparison. The reason that I need to pass in a builder is because my schema contains possibilities for a graph, and the only way to represent this is with SchemaBuilder, as ConnectSchema is not mutable.